Fix the truncating of UTF8 characters
When data is too long to fit in the columns of a database table, it is
cut such as edit comments. However with UTF8, some characters may be
several bytes long and cutting them in the middle produce invalid SQL
files.

For example
https://dumps.wikimedia.org/enwiki/20210901/enwiki-20210901-langlinks.sql.gz
shows the issue if you do:

  zcat enwiki-20210901-langlinks.sql.gz | isutf8

On linux you get:

  (standard input): line 139, char 892171, byte 103389006:
  After a first byte of E0, expecting a 3nd byte between 80 and BF.

Some patches were send to T15721:
* 'Trim replacement character from rev_comment' merged as dddb2224
* Ieff7eba1523c2d89465141fe59a4aef1ca42c1f4 was never merged and was
  intended to address the UTF-8 truncating issue.

Bug: T15721
Change-Id: I9ad22ca198d19a65005010e515ed63a6ab0738d5